[SPARK-37424][PYTHON][MLLIB] Inline typehints for pyspark.mllib.random
### What changes were proposed in this pull request?

This PR migrates type `pyspark.mllib.random` annotations from stub file to inline type hints.

### Why are the changes needed?

Part of ongoing migration of type hints.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing tests.

Closes #35576 from zero323/SPARK-37424.

Authored-by: zero323 <mszymkiewicz@gmail.com>
